+++ include/encoder.h
@@ -8,6 +8,6 @@
 #include "structs.h"
 
 void getSavedName(char *name, char *buff);
-int readPpm(FILE* f, int raw[3][PIX_LEN]);
-int writePpm(FILE * f, int sub[3][PIX_LEN/16]);
-void encodeNsend(char * name, int raw[3][PIX_LEN], area_t dims, int dim);
+int readPpm(FILE* f, uint8_t raw[3][PIX_LEN]);
+int writePpm(FILE * f, uint8_t sub[3][PIX_LEN/16]);
+void encodeNsend(char * name, uint8_t raw[3][PIX_LEN], area_t dims, int dim);
